+++ lib/file_saver_web.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:developer';
 // In order to *not* need this ignore, consider extracting the "web" version
 // of your plugin as a separate package, instead of inlining it in the same
 // package as the core of your plugin.
@@ -54,7 +53,7 @@ class FileSaverWeb {
       document.body!.children.remove(anchor);
       success = true;
     } catch (e) {
-      log(e.toString());
+      rethrow;
     }
     return success;
   }