+++ android/src/main/kotlin/com/incrediblezayed/file_saver/Dialog.kt
@@ -60,20 +60,13 @@ class Dialog(private val activity: Activity) : PluginRegistry.ActivityResultList
         this.fileName = fileName
         val intent =
             Intent(Intent.ACTION_CREATE_DOCUMENT)
-        intent.putExtra(Intent.EXTRA_LOCAL_ONLY, true)
         intent.addCategory(Intent.CATEGORY_OPENABLE)
         intent.putExtra(Intent.EXTRA_TITLE, "$fileName.$ext")
-        intent.putExtra(Intent.EXTRA_MIME_TYPES, type)
         intent.putExtra(
             DocumentsContract.EXTRA_INITIAL_URI,
             Environment.getExternalStorageDirectory().path
         )
         intent.type = type
-        intent.flags = (Intent.FLAG_GRANT_READ_URI_PERMISSION
-                or Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-                or Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION
-                or Intent.FLAG_GRANT_PREFIX_URI_PERMISSION
-                )
         activity.startActivityForResult(intent, SAVE_FILE)
     }